apps/preflight-api/requirements.txt
fastapi>=0.112.0
uvicorn[standard]>=0.30.6
SQLAlchemy>=2.0.34
alembic>=1.13.2
psycopg>=3.2
asyncpg>=0.29.0
python-dotenv>=1.0.1
ruff>=0.6.4
pytest>=8.3.2
pytest-asyncio>=0.24.0
httpx>=0.27.0
openai>=1.50.0
anthropic>=0.36.0
tenacity>=9.0.0
pydantic-settings>=2.5.0
PyJWT>=2.8.0
redis>=5.0.0
